python/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "SALT Web Manager"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
python-multipart = "^0.0.5"
Jinja2 = "^2.11.3"
aiofiles = "^0.6.0"
dsnparse = "^0.1.15"
aiomysql = "^0.0.21"
astropy = "^4.2.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-print = "^4.0.0"
tox = "^3.21.4"
isort = "^5.7.0"
mypy = "^0.800"
pytest-cov = "^2.11.1"
bandit = "^1.7.0"
mkdocs-material = "^6.2.7"
requests = "^2.25.1"
pydantic = {extras = ["dotenv"], version = "^1.8.1"}
pytest-asyncio = "^0.14.0"
mkdocs-macros-plugin = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
